fix: switched to pandoc for markdown + crazy css
because was losing the double backslashes in mathjax,
using cmark/marko

css slightly crazy.. .
got the mathjax expanding nicely.
but also expanded the code outputs..
tbh not sure if that's the way to go.
it looks good for the fourier plots,
but when there's just a bit of text,
it doesn't look great.